requirements.txt
Django>=6.0.0
djangorestframework>=3.14.0
PySide6>=6.6.0
requests>=2.31.0
mitmproxy>=10.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
waybackpy>=3.0.6
clamd>=1.0.2
psycopg2-binary>=2.8
PyQt6==6.7.0
PyQt6-WebEngine==6.7.0
selenium>=4.15.0
webdriver-manager>=4.0.1
fake-useragent>=1.4.0
Pillow>=10.1.0
gunicorn>=21.2.0
whitenoise>=6.6.0
celery>=5.3.0
redis>=5.0.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
PyYAML>=6.0
playwright>=1.40.0
Flask>=3.0.0
imageio>=2.33.0
imageio-ffmpeg>=0.4.9
jsbeautifier>=1.14.0
# Malware analysis dependencies (optional - installed with graceful fallback)
pefile>=2023.2.7
yara-python>=4.3.0
scikit-learn>=1.3.0
joblib>=1.3.0
pyelftools>=0.30
androguard>=3.3.5
olefile>=0.47
reportlab>=4.0.0
scapy>=2.5.0
stix2>=3.0.0
# OSINT engine dependencies (added for enhanced discover app)
dnspython>=2.4.0
python-whois>=0.9.0
cryptography>=41.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
openpyxl>=3.1.0
aiohttp>=3.9.0
networkx>=3.2
